Add opening hours form actions component
