+++ src/components/opening-hours-form/OpeningHoursForm.scss
@@ -28,41 +28,6 @@
   }
 }
 
-.opening-hours-form__actions-container {
-  background: #fff;
-  bottom: 0;
-  box-shadow: 0 -2px 10px 0 rgba(212, 212, 212, 0.2);
-  border-top: 1px solid var(--hauki-content-border-color);
-  left: 0;
-  position: fixed;
-  right: 0;
-  z-index: 3;
-}
-
-.opening-hours-form__actions {
-  align-items: center;
-  box-sizing: border-box;
-  display: flex;
-  gap: 1rem;
-  justify-content: center;
-  margin: 0 auto;
-  max-width: $container-width-xl;
-  padding-bottom: 1rem;
-  padding-top: 1rem;
-}
-
-@media screen and (min-width: $breakpoint-m) {
-  .opening-hours-form__actions {
-    gap: 1.5rem;
-  }
-}
-
-@media screen and (min-width: $breakpoint-l) {
-  .opening-hours-form__actions {
-    justify-content: left;
-  }
-}
-
 .opening-hour-forms-required-help-text {
   font-size: 1rem;
   margin: 0;
