+++ src/components/opening-hours-form/OpeningHoursFormActions.scss
@@ -1,0 +1,36 @@
+@import 'node_modules/hds-design-tokens/lib/all.scss';
+
+.opening-hours-form-actions-container {
+  background: #fff;
+  bottom: 0;
+  box-shadow: 0 -2px 10px 0 rgba(212, 212, 212, 0.2);
+  border-top: 1px solid var(--hauki-content-border-color);
+  left: 0;
+  position: fixed;
+  right: 0;
+  z-index: 3;
+}
+
+.opening-hours-form-actions {
+  align-items: center;
+  box-sizing: border-box;
+  display: flex;
+  gap: 1rem;
+  justify-content: center;
+  margin: 0 auto;
+  max-width: $container-width-xl;
+  padding-bottom: 1rem;
+  padding-top: 1rem;
+}
+
+@media screen and (min-width: $breakpoint-m) {
+  .opening-hours-form-actions {
+    gap: 1.5rem;
+  }
+}
+
+@media screen and (min-width: $breakpoint-l) {
+  .opening-hours-form-actions {
+    justify-content: left;
+  }
+}
